upgrade-test Move classes of com.liferay.portal.upgrade.v7_0_0 package to module
